backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.25
pandas==2.2.0
sentence-transformers==2.3.1
faiss-cpu==1.7.4
python-multipart==0.0.6
openpyxl==3.1.2
pydantic==2.6.1
aiofiles==23.2.1
numpy==1.26.4
scikit-learn==1.4.0
